Cleanup redundant modules (#140)
* Rename kwarg

* Move Intrinsics to separate file

* Deprecate useless module

* Deprecate animations

* Deprecate dcm2nii
