+++ src/xvr/io/__init__.py
@@ -1,3 +1,4 @@
+from .intrinsics import Intrinsics
 from .xray import read_xray
 
-__all__ = ["read_xray"]
+__all__ = ["read_xray", "Intrinsics"]
