+++ src/xvr/io/xray.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass, fields
 from pathlib import Path
 from typing import Callable
 
@@ -8,24 +7,7 @@ from jaxtyping import Float
 from pydicom import dcmread
 from torchvision.transforms.functional import center_crop
 
-
-@dataclass
-class Intrinsics:
-    sdd: float
-    delx: float
-    dely: float
-    x0: float
-    y0: float
-
-    def __post_init__(self):
-        for f in fields(self):
-            setattr(self, f.name, float(getattr(self, f.name)))
-
-    def keys(self):
-        return [f.name for f in fields(self)]
-
-    def __getitem__(self, key):
-        return getattr(self, key)
+from .intrinsics import Intrinsics
 
 
 def read_xray(
