--- src/xvr/renderer/__init__.py
@@ -1,3 +1,0 @@
-from .load import initialize_drr
-
-__all__ = ["initialize_drr"]
