--- src/xvr/visualization/__init__.py
@@ -1,4 +1,0 @@
-from .animate import animate
-from .viz2d import plot_registration
-
-__all__ = ["animate", "plot_registration"]
